feat: show profile settings in General for a single profile, link to Profiles settings otherwise
